test: Rewrited the rooms.helper file in typescript
This should give better intelisense when writing api tests
