+++ apps/meteor/tests/end-to-end/api/07-incoming-integrations.js
@@ -4,7 +4,7 @@ import { after, before, describe, it } from 'mocha';
 import { getCredentials, api, request, credentials } from '../../data/api-data.js';
 import { createIntegration, removeIntegration } from '../../data/integration.helper';
 import { updatePermission } from '../../data/permissions.helper';
-import { createRoom, deleteRoom } from '../../data/rooms.helper.js';
+import { createRoom, deleteRoom } from '../../data/rooms.helper';
 import { password } from '../../data/user';
 import { createUser, deleteUser, login } from '../../data/users.helper';
